Use the Gradle worker api to run the generators
At the moment the generators are loaded directly into the JVM of the
Gradle daemon. This can cause metaspace problems and makes parallel
task execution impossible. Furtheremore, this introduces classloader
leaks that are hard to fix. Using the Gradle worker api to execute the
generators in separate processes should resolve these issues.