Set visibility of configurations to false
Artifacts of visible configurations are attached to the archives
configuration and thus are built when executing the assemble task.
However, the source and model zipping tasks are only required in the
rare case that we intend to publish sources and models.
Therefore we set the visibility of all our configurations to false.