Update to Gradle version 7.6
Increases the minimum required Gradle version to 7.6 because that's the
first version where worker classpaths are cached by Gradle itself, see
[gradle/gradle#21475](https://github.com/gradle/gradle/pull/21475).
This will enable us to remove the caching of the worker classpath that
we do ourselves.

Also, all Gradle plugins used in the build are updated to the latest
version and various deprecation warnings in the build are fixed.